modify : 2nd merge feedback and convention
