+++ src/pages/Admin/components/AdminContainer/components/DatePickerComponent.tsx
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
 import DatePicker from 'react-datepicker';
+import styled from 'styled-components';
 import 'react-datepicker/dist/react-datepicker.css';
 import { ko } from 'date-fns/esm/locale';
 import Form from 'react-bootstrap/Form';
