+++ src/pages/Admin/components/Modal/NoticeModal/components/NoticeBox.tsx
